requirements.txt
requests
feedparser
python-dotenv
beautifulsoup4